+++ src/olim8_mp0.test.cpp
@@ -45,7 +45,7 @@ BOOST_AUTO_TEST_CASE (neighboring_values_are_correct) {
 }
 
 BOOST_AUTO_TEST_CASE (origin_test) {
-  size_t M = 5, N = 5;
+  int M = 5, N = 5;
   double h = 0.1, x0 = h*(N - 1)/2., y0 = h*(M - 1)/2.;
   olim8_mp0 m {M, N, h, default_speed_func, x0, y0};
   m.add_boundary_node(2, 2);
@@ -60,12 +60,12 @@ BOOST_AUTO_TEST_CASE (origin_test) {
 }
 
 BOOST_AUTO_TEST_CASE (sf1_single_row_test) {
-  size_t N = 1001;
+  int N = 1001;
   double h = 1.0/(N - 1);
   olim8_mp0 m {1, N, h, s1};
   m.add_boundary_node(0, 0);
   m.run();
-  for (size_t j = N - 10; j < N; ++j) {
+  for (int j = N - 10; j < N; ++j) {
     double U = m.get_value(0, j);
     double u = f1(h*j, 0);
     printf("%g\n", fabs(u - U)/fabs(u));
@@ -74,13 +74,13 @@ BOOST_AUTO_TEST_CASE (sf1_single_row_test) {
 }
 
 BOOST_AUTO_TEST_CASE (sf1_test) {
-  size_t M = 101, N = M;
+  int M = 101, N = M;
   double h = 1.0/(M - 1);
   olim8_mp0 m {M, N, h, s1};
   m.add_boundary_node(0, 0);
   m.run();
-  for (size_t i = M - 3; i < M; ++i) {
-    for (size_t j = N - 3; j < N; ++j) {
+  for (int i = M - 3; i < M; ++i) {
+    for (int j = N - 3; j < N; ++j) {
       double U = m.get_value(i, j);
       double u = f1(h*j, h*i);
       printf("%g\n", fabs(u - U)/fabs(u));
